Fixes #2459. Remove last traces of the users_registration table from tables.php and the rest of the code.
Most references to users_registration were in comments, however there were some references to the table in code. In Users/lib/Users/Api/Registration.php, the countAll function had an unused reference to the table. In Users/lib/Users/Api/User.php, the mailUname function was still using the table to determine if an email address was used, and would have resulted in a failure. Finally, in Users/tables.php the users_registration table definition was removed from the Users_tables_for_210 function.