Detect mailer failure and ensure consistent users_verifychg state. Display error icon if verification not sent and it was supposed to have been. Fixes #2473.
If the Mailer returns false when sending an e-mail verification code, ensure that the users_verifychg record is removed from the database so that the pending registration record does not indicate that the message was sent.

If an email verification message was supposed to have been sent already, given the state of the pending registration, then display an error icon to indicate to the administrator that he must resend the message.