Move status-based failed login messages to Users_Controller_User class. Fixes #2480.
The code that provided status-based failed login messages was in UserUtil (and probably still needs to be there), but because of the change in the way pending registrations are being handled, the login process from the Users module never got to that point. Move the status checks to the Users module.

Change the users_user_registrationfinished.tpl to a more generic users_user_displaystatus.tpl, so it can be used with both the login process and the registration process.