Fixed not working background image in 'Input'-button for IE-browsers (refs #2504) and added the first CSS3PIE styles (refs #2509).
Not yet finished. This is rather a proof of concept and I would really like to find another solution. We have now a lot of CSS rules our buttons in styles/core.css. This is very complex and difficult. If we only use the button tag, then this would not be a problem. But IE has major problems with more than a button-tag per page. For that we have a javascript-fix by Jusuff. But if JS is disabled, then the page can not be used. For this reason we have, unfortunately, still buttons based on the input tags. And unfortunately, it is quite complicated to make them look equal in all browsers. Such is life. :-)