Add dash (-) as a legal character in a user name. Fixes #2513.
In 1.2.4 and in 1.3.X sometime prior to this commit, the definition of a valid user name was refined to include only numbers, letters, the underscore and period. After the release of 1.2.4, site operators noted that they had users with user names that include a dash. This re-enables dash as a legal character in user names.