fixed wrong display of search options when modifying a search box, closes #2522
When no module was selected in search options, the block displayed them all as checked (instead of the opposite).
This was because of a variable $vars['active'] which every module's search logic is checking. When no module was selected in search options, this variable was not set, causing the modules to 'activate' their search boxes.
Problem was fixed by instantiating the variable $vars['active'] as an empty array (if it wasn't set before).