Altered the EventManager to allow weighting of event handlers refs #2605.
Currently in the event manager it's not possible to influence the order in which event handlers are executed. We discussed this in the summer and came to the conclusion that it could still be difficult to be 100% of execution order because there is no way to be sure of what would change the load order in the course of the execution cycle.

This change allows us to sort by weight when attaching new handlers. Then handlers can be registered and sorted according to weight.

Made some small optimisations in the EventManager in the process.

Also made some changes to the Event object to allow return of $this for convenience.