New template solution for the sorting of table columns. Now easier to understand. Refs #2579
I've tried plugins and modifiers and couldn't find any more flexible solution that we can use anywhere. The problem is that we need the output for the CSS class for 'SortBy'- and 'SortDirection'-parameters . Thus, we would have to call a modifier three times. And I think that is easier to use a simple condition in the templates.