Rework modifier logic to properly use DataUtil::formatForStore and DataUtil::formatForDisplay. Fixes #2204
The modifier was using DataUtil::formatForDisplay inappropriately for data it was using to look up the user record. Fixed the logic to use DataUtil::formatForStore where needed. Also fixed so that the link always uses uid to look up the profile record, and never uname. This prevents problems where the uname has characters incompatible with URLs, but where replacing the characters with entities will cause the query in Profile to fail.