Allow guest user to return back attributes refs #2462
This is getting far too dirty.  UID=0 is hard coded in so many places, it's not possible to return the real UID of the guest user so we have to fake the case.  The other problem is that because the guest user is not a real user, it doesn't actually have any extended attributes like Real Name. Given we can only really get back the uname I'm begining to things this is a complete waste of time.

Overall it might be necessary for modules to just test for Guest user and than rely on other settings, and then make it's own assumptions from there and just forget the rest of it. The only real way out of this is to have a proper user object which is not possible in this milestone.