change logic in new user creation template, refs #2673
The logic in the template autosets "setpass" (a radio box stating if you want to set the users pass or not) to "no", this makes sense.
however the logic also hides the "setpass" radio's by the same logic, so without JS this value would be set to "no" and you would never
see the option to change it. This has been changed so you should always see the "setpass" options when not using JS.