sort the providers using the Hook UI, refs #2604
after changeset c25c9268901f2d7ab9e8dfdda69708d13a587c26 the actual sorting is not done on the providers but their respective positions.
This changeset implements the new way of sorting, while keeping the actual implementation transparent to the user.
What this means is that the user still does the sorting on the providers, and under the hood the sorting is done on the areas.