Check to see if $item.active is set before checking if it is true. Fixes #2704
The original code checked active with a simple if ({if $item.active ...}). If the item was not active, the active element of $item was never defined, and therefore generated an "Undefined index" notice. The notice interrupted the output, thus sending the remainder of the HTML code out as normal output.