Reorder user name, e-mail and password fields to support user name / password autofill Closes #2732
Modern browsers, such as FireFox, remember user names and password for various web sites the user visits (if the browser is configured to do so). In order to find the user name field, the browser finds the password field (through inspecting various attributes), and then assumes that the user name is in the input field immediately preceding the password on the form.