Gettext incorrectly initialized for modules if using custom short urls closes #2740
If custom shorturl decoders are used, they invoke modutil::apifunc which will then bind domain, but this happens before the language system determines the language (which has to be taken from the url).  The better solution would be to move decoders to an event since doesn't require the standard initialisation process.