Don't assume user name and password, but also, if there are no modules capable of authentication, then force user name and password. Refs #2709
In rare cases ModUtil::getModulesCapableOf('authentication') may return no modules at all. If no additional authentication modules are installed or activated, and the Users module is in a state other than 'active' (e.g., a new version was installed and is awaiting upgrade), then there will be no modules capable of authentication. We know that the Users module is installed, and we also know that the admin user has a user name and password (unless the site admin did something silly), so log a critical error and force the authentication module to be the Users module. Hopefully in it's other-than-active state the admin user can still log in.