Localized category name is marked mandatory now. Closes #2749
In the case that the localized category name isn't entered, it is copied now from the name.

I think we all validation certainly could easily be resolved with FormUtil. But the currently used validation of the category name does not work either. For this reason I will now pursue this direction any further. Perhaps RNG can take a look on this: After saving a blank page (without a category name) the user sees no error messages.