major refactor of the administration interface for hooks -- NEEDS REINSTALL -, closes #2604
The ui for administering hooks has completely changed.
Now it is possible to attach a specific area of a provider module to a specific area of a subscriber module.
It is also possible to reorder provider areas that are attached to the same subscriber area.

The ui is now accessible from the menu item link "Hooks", where the user is presented with the configuration page. This page is the same for both provider modules and subscriber modules. The system checks to see the capabilities and shows the exact section accordingly.