Shortened the installation process by removing unnecessary installer check screen.
The less steps (and clicks) required for an installation, the easier it
is for the user.  There is no point displaying a page of passed checks.
The installer will now silently check the install requirements and if they
pass, it will silently pass onto the next screen.  If the checks fail, the
user will be presented with a list of problems to fix.