Clarify language refs #2785
This menu option specified the default route.  That is, when a URL does not
contain the module name as part of the routing, where should Zikula send
requests to. For example, if we set this to News module then /2011/1/29/news_introduction_article/ will be decoded and sent to the News module.

If no default route is set, Zikula will be unable to route the request anywhere and
will issue (correctly) a 404.  This settings also affects short URL generation.