Changed "not installed" workflow refs #2801 refs #2807
In light of the various usability changes I've refactored a few things to accommodate
for both these ticket.  With this commit notinstalled.tpl template is now fully
gettexted and resides in installer.  Localised versions can be created by adding
config/installer.ini (language = xx).

There was no possibility of adding a language selector to the notinstalled.tpl
template because the language system had not initialised in the previous version
and it would not be possible to have a translated language selector.  This is now
possible although does not feature in this commit.