Correct outputfilter.pagevars.php interference with function.zdebug.php
The pagevars output filter was using str_replace to perform its function. This is fine if zdebug or something else is not trying to open up another pop-up window with Javascript. If zdebug (or something else) is trying to do so, then there will be more than one </head> tag: one for the main page being displayed and another for the console window (or whatever) being opened in the pop-up.

Change checks for more than one </head> and acts appropriately to restrict pagevar's actions to the </head> of the main page.

Also, move the nasty bits of the zdebug javascript into a .js file, and update the template to use that .js file.