load stylesheets properly in blocks Extmenu and Menu
the stylesheets were not loaded correctly if render caching was enabled.
This was due to the fact that PageUtil::addVar was used AFTER the checking of whether the cached template should be used or not.
So if a cached template was used, the stylesheet was never loaded.
Fixed the issue by moving the call to PageUtil::addVar before the cache check.