special characters were not replaced with given permalink replacements
there was a preg_replace function which was used to replace spaces and other chars in the permalink. however this function was called before the substitution of permasearch/permareplace characters. this caused problems with umlauts and possible other characters.

fixed the problem by moving the preg_replace function AFTER the substitution of permasearch/permareplace characters.

see http://code.zikula.org/core/ticket/2580 for more details