Ensure that events in Users happen appropriately. Include redirecturl in registration and login succeeded events. Fixes #2888, #2897
To change where a user is redirected to, intercept the event with the redirecturl argument and change its value. The process will then redirect to the desired place.

There is no user.login.failed event. The user simply goes back to the login page. Do we need one?

Ensured that the registration process has a proper set of events, and that a registration record has the appropriate create, update and delete events.