Minor touchs to the module links display
Removed the left border for the first link because its already on the left border, and reduced the padding to look consistent.
Also makred all the links with classes with the additional z-iconlink class, and reduced the space between the icon and the
separator with a negative left margin for them.