Simplify view.init and theme.init events and applied some naming standards and cleanup.
Everything in the event arguments can be obtained directly from the object.
Removed unused params in constructor of theme (scanned use in core and these are orphaned).
Applied some naming standards in the constructors: $theme might actually mean a theme object
as opposed to $themeName which is quite specifically a name.