Pager plugin 'optimize' parameter enabled by default.
The pagercss.tpl template expects optimized links already, and makes no sense
to generate te complete list by default causing performance issues because modurl.
Review of the math used calculating the side limits when 'maxpages' is used.

Also, the generation of the links was changed to not include the first and last links
always, because it's not needed and those ones are available by default.