New template block function: pageaddvarblock. Allows inline scripts, inline styles, etc. to be placed in head ('rawtext') or 'footer' (just before </body>).
This template block function works similarly to {pageaddvar}, however only the 'rawtext' and 'footer' page variables are allowed. 'rawtext' is the default if no page var name is specified. This block function will place the contents of the block in either the head section of the page (if name is 'rawtext' or is not specified), or in the area just prior to the closing body tag (if name is 'footer'). It can be used to (among other things) place inline script blocks and inline style blocks specified in the template in a more appropriate place within the page, rather than leaving them in the flow of the HTML markup.

For example, if the script within an inline script block will run when the DOM is loaded, it does not need to be in the flow of the HTML markup. Instead, it should be placed in the head section of the page. Surrounding the inline script block with {pageaddvarblock}...{/pageaddvarblock} (or explicitly, {pageaddvarblock name='rawtext'}...{/pageaddvarblock}) will place it there as if it had been placed there by PageUtil::adVar('rawtext', $content). (In fact that is exactly what happens within the block function.) Likewise, inline style blocks should be placed in the head section. Script blocks that cannot be placed in the head section, but that do not emit anything to the document's markup or modify any form controls in immediate mode, can be moved to the end of the document's markup in order to (very slightly, almost insignificantly) improve the document's rendering. Surround these blocks with {pageaddvarblock name='footer'}...{/pageaddvarblock}.