Deprecate the page variable 'rawtext' and replace it with 'header'.
The page variable name 'header' is more intuitive and more in line with the existing 'footer' variable name.

If the system is operating in legacy mode, then 'rawtext' is still accepted, and converted to 'header' on the fly. If not operating in legacy mode, then 'rawtext' is treated like any other randomly named page variable and is not rendered to the page.