Added support to nested sets on Zikula.Tree. Refs #2998
Added objid parameter to specify the id field to use.
Added nestedSet parameter (default: false)
Added renderRoot parameter to enable or not the inclusion of the root node on the output.
Added fixedParent class to mark links and treat them as parents always.

Items can have direct nodes that are not indexed as the others.

Added identation of the Zikula.Tree HTML output making it readable
and some coding standards fixes.

Categories sync to the new parameters