Fixed invalid short-url generation in edgecase refs #3002
There is an edgecase when using ModUtil::url() as an extra argument in the $args
to add more values to the URL when in short URL mode (like a redirect url).
ModUtil::url() will generate directory based urls with a trailing slash.  Because
the argument may be wrapped in encodeurl() this will mung the directory based URL with
and encoded slash as %2F and break valid URL scheme.