Mark the Zikula.UI.Tabs  content containers with a CSS class (z-tabs-content) by default
Added default borders and stylñes for them, to be consistent with the expected layout.
Also added a info box with the ways to call Zikula.UI in the examples page,
and a CSS line to avoid the Firefox dotted border on the active tab link.

The question is:
Was the double right border for unactive-tab:hover on purpose?