Made possible to call $themeEngine->load_config to alter the theme output from the controller.
It was not enough to be able to add the Admin module plugins and the admintitle.outputfilter
but also to be able to reload the config of the theme with a different value, in this case,
a forced $themeEngine->type = admin.

Then, we can have a controller Import, Edit, etc. with this base postInitialize method:
public function postInitialize()
{
    $serviceManager = ServiceUtil::getManager();
    $themeInstance  = $serviceManager->getService('zikula.theme');
    $themeInstance->type = 'admin';
    $themeInstance->load_config();

    $this->view->addPluginDir('system/Admin/templates/plugins');
    $this->view->load_filter('output', 'admintitle');
}

And render its output as if it was the Admin Controller.

I've moved out some calls to not re-do something that can be done outside load_config one time only.