Add an API requirement to the Authentication API for user interface level authentication information validation. Refs #3031
It is not possible to provide generic authentication of the user-entered authentication information (log-in credentials) at the user interface level, because each authentication method defined by an authentication module defines its own requirements for the credentials entered (if any). Add a requirement to the Controller component of the Authentication API to allow for the validation of these credentials by the authentication module.