Default block not choosen by default to avoid mistakes on Block Create Form. Refs #3034
Validation of an existing one in the create method-
Extension of the newblock template to allow default values:
when a module creation fails the user does not lose the input submited.

`ModUtil::url` urlencode fix for array values as arguments.
also an extension receiving an array of arrays as arguments.
It's needed for the new block creation error returnurl.