Register sensible core plugins as non-cacheable when Render cache is enabled. Refs #3030
This avoid the need of have inserts for everything to work under cache, and adds the possibility
to cache the parameters of them. The registered plugins are the ones related wuth the pagevars
or user-based information/processing.

Zikula_View_Resource is in charge of their lazy load when cache is enabled.