Fix collapsing button area when checking registrations.
When a user clicks on the button to check his registration entries the area containing the button and message collapse to an empty fieldset and the ajax progress indicator is not displayed. Corrected the style and class information so that this does not happen and so that the progress indicator is displayed.