On field or validation errors, set the next state to redisplay the registration form. Refs #3057
This is the primary cause of the infinite loop mentioned in #3057. If there were validation errors, then the function was not redirecting to redisplay the form. It was getting lost in the (previous) state machine, re-executing the 'register' state over and over. With the restructuring of the state machine, and the addition of the state set on error, the infinite loop should be solved.

Also, fix the template args errorFields and errorMessages being sent from the registration method.