Allow embedded underscores in template function parameters when prefiltering for double-underscore gettext replacements. Refs #3061
The previous regular expression did not allow for parameters such as `__icon_alt` or `__icon_title`. This updated regex allows for embedded underscores. There must be at least one alpha-numeric character between the double-underscore and an embedded underscore. For example, the following are all legal:

    __icon_alt
    __9_west
    __A_title
    __title_
    __12_

The following is not legal:

    ___title (note three underscores in a row)