Allow for an icon to appear along with the helplink's specified text. This enables further formatting into a button. Closes #3061
The ability to fully specify an icon to include along with the helplink link text has been added by prefixing all {icon} function parameters with 'icon_' and adding them to the {helplink} function. The icon can appear on either the left side (default) of the text, or the right side. It is separated from the text with a non-breaking space to prevent unwanted mid-link breaks.

The icon can simply be added on its own, becoming part of a "normal" link, or if the {helplink} appears inside a div or span that has a class attribute containing the value 'z-buttons', then the core stylesheet will format the link as a button. Nothing in the {helplink} function needs to be specified to format the link as a button, and the icon is not necessary (but is customary) for the link to appear as a button. Simply wrapping it in a div or span with the appropriate z-buttons class is all that is needed.