provide more user-friendly error response when trying to view Theme module as user ad theme-switching is disabled. refs #3084
the extmenu has an option to "Display links to installed modules". This obviously provides a link to the Theme module when it would not be otherwise available. In this edge-case, a more appropriate error message should be provided and then redirect the user to a suitable destination instead of displaying 'page not found' error.