Simplification of the admin templates. Refs #3088
Now it is no longer necessary to open two separate areas (z-adminbox and z-admincontainer) for the admin area. Through simple plug-ins now, a box is drawn around the module output. Previously, we had (now and then) the problem that a gap between the "z-adminbox" and "z-admincontainer" was created by CSS paddings. This can not happen anymore.

In addition, the code is leaner and a module developer must not take care on the exact order of the HTML structure. No more unnecessary lines in admin templates.