add support for Doctrine 2 categories in plugin assignedcategorieslist
If you are using Doctrine 2, you must explicitly pass the categories as parameter to the plugin. Also, you must set a parameter doctrine2 to true. This is necessary because for Doctrine 2, the categories are assigned to a custom variable defined at the Entity, so it's not known to the plugin.

Support for Doctrine 1 and DBUtil remains as before.