Fix response workflow.
Removed support for boolean controller response, must return a Response object.
Use PlainResponse to override the theme.

Fixed some E_NOTICES in Users while I was at it.
Removed distinction between meta and WCAG redirects, this is nonesense, the right way is to
issue a redirect header with a meta refresh which is done by the RedirectResponse already.