refactor Settings/modfiyconfig to twig
This needs a lot more work to be completely refactored.
We need twig extensions to replace the smarty plugins. In this specific template we need the following:
ajaxheader
adminheader
adminfooter
img
icon
button
modurl
insert
html_select_modules

On a sidenote, how do we access the $modvars? This should be written in the Twig refactoring guide.