fixes #621, added fileupload input styling Webkit
webkit browsers have a different way of showing input file upload fields. 
This is messed up with the current formrow styling with display: block. 
This is fixed here, Firefox still works ok. Not tested in all browsers yet.

Styling of the webkit button is chosen the same as the regular buttons in core.css.