Allow new type of theming to co-exist with old.
The better way of theming will be to use template inheritance
features of Twig. As such a controller simply includes a base
template (which can be overriden by the theme system).

This is a simple hack which prevents the old theme from rendering
if module was found using routing (which only new styles can).