Updated external dynamic links menutree with Clip
| Q             | A
| ------------- | ---
| Bug fix?      | no
| New feature?  | yes
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | -
| Fixed tickets | -
| Refs tickets  | -
| License       | MIT
| Doc PR        | -

Clip publications are added to the available list of links one can dynamically add/generate in a menutree menu. I have this active currently already in Zikula Core 1.3.6 with the current Clip from Github. 

Of course the future of Clip is uncertain. So maybe this can be discussed before pulled.