[1.4.x] Updated DBUtil for OrderBy and table prefix
See https://github.com/zikula/core/pull/1924 
this is the same PR for 1.4.x

ZWebstore uses this function and ordering of products is not working without this small update. It's an extra check for table prefix in the column checking. Not checked in 1.4.x. yet.

I have checked with other modules using DBUtil as well (e.g. Content), and there no problems arise due to this small change.

| Q             | A
| ------------- | ---
| Bug fix?      | yes
| New feature?  | no
| BC breaks?    | no
| Deprecations? | no
| Tests pass?   | -
| Fixed tickets | -
| Refs tickets  | -
| License       | MIT
| Doc PR        | -